+++ pixeltable/catalog/catalog.py
@@ -1053,23 +1053,15 @@ class Catalog:
         return Path.parse('/'.join(names), allow_empty_path=True, allow_system_path=True)
 
     def _table_error_counts(self) -> dict[UUID, int]:
-        """Returns map from table id to the sum of num_excs across that table's versions.
-
-        Sums tableversions.md -> update_status -> row_count_stats / cascade_row_count_stats -> num_excs
-        for each row, grouped by tbl_id.
-        """
-
-        # JSONB path mirrors UpdateStatus.row_count_stats / cascade_row_count_stats -> RowCountStats.num_excs
-        # (see pixeltable/catalog/update_status.py). If those dataclass fields are renamed, this query
-        # silently returns 0.
-        stmt = sql.text("""
-            SELECT tbl_id, COALESCE(SUM(
-                COALESCE((md -> 'update_status' -> 'row_count_stats' ->> 'num_excs')::int, 0)
-              + COALESCE((md -> 'update_status' -> 'cascade_row_count_stats' ->> 'num_excs')::int, 0)
-            ), 0) AS errors
-            FROM tableversions
-            GROUP BY tbl_id
-        """)
+        """Returns map from table id to the sum of num_excs across that table's versions."""
+        md = schema.TableVersion.md
+        update_status = md['update_status']
+        row_count_excs = sql.func.coalesce(update_status['row_count_stats']['num_excs'].astext.cast(sql.Integer), 0)
+        cascade_row_count_excs = sql.func.coalesce(
+            update_status['cascade_row_count_stats']['num_excs'].astext.cast(sql.Integer), 0
+        )
+        errors = sql.func.coalesce(sql.func.sum(row_count_excs + cascade_row_count_excs), 0).label('errors')
+        stmt = sql.select(schema.TableVersion.tbl_id, errors).group_by(schema.TableVersion.tbl_id)
         rows = get_runtime().conn.execute(stmt).all()
         return {r.tbl_id: r.errors for r in rows}
